spinner.go: Add support for windows
